update generated code

+++ src/common/UML4CPPProfile/src_gen/UML4CPPProfile/UML4CPPProfilePackage.hpp
@@ -91,35 +91,35 @@ namespace UML4CPPProfile
 
 			//Static IDs
 				//Class UML4CPPProfile::DoNotGenerate
-				static const long long DONOTGENERATE_CLASS = 173581214;
+				static const unsigned long DONOTGENERATE_CLASS = 173581214;
 				static const unsigned int DONOTGENERATE_CLASS_FEATURE_COUNT = 1;
 				static const unsigned int DONOTGENERATE_CLASS_OPERATION_COUNT = 0;
 				//Class UML4CPPProfile::ExternalLibrary
-				static const long long EXTERNALLIBRARY_CLASS = 744219230;
+				static const unsigned long EXTERNALLIBRARY_CLASS = 744219230;
 				static const unsigned int EXTERNALLIBRARY_CLASS_FEATURE_COUNT = 4;
 				static const unsigned int EXTERNALLIBRARY_CLASS_OPERATION_COUNT = 0;
 				//Class UML4CPPProfile::GetterName
-				static const long long GETTERNAME_CLASS = 1711019851;
+				static const unsigned long GETTERNAME_CLASS = 1711019851;
 				static const unsigned int GETTERNAME_CLASS_FEATURE_COUNT = 2;
 				static const unsigned int GETTERNAME_CLASS_OPERATION_COUNT = 0;
 				//Class UML4CPPProfile::MainBehavior
-				static const long long MAINBEHAVIOR_CLASS = 1527007467;
+				static const unsigned long MAINBEHAVIOR_CLASS = 1527007467;
 				static const unsigned int MAINBEHAVIOR_CLASS_FEATURE_COUNT = 1;
 				static const unsigned int MAINBEHAVIOR_CLASS_OPERATION_COUNT = 0;
 				//Class UML4CPPProfile::NonExecutable
-				static const long long NONEXECUTABLE_CLASS = 1641344257;
+				static const unsigned long NONEXECUTABLE_CLASS = 1641344257;
 				static const unsigned int NONEXECUTABLE_CLASS_FEATURE_COUNT = 1;
 				static const unsigned int NONEXECUTABLE_CLASS_OPERATION_COUNT = 0;
 				//Class UML4CPPProfile::SetterName
-				static const long long SETTERNAME_CLASS = 103590585;
+				static const unsigned long SETTERNAME_CLASS = 103590585;
 				static const unsigned int SETTERNAME_CLASS_FEATURE_COUNT = 2;
 				static const unsigned int SETTERNAME_CLASS_OPERATION_COUNT = 0;
 				//Class UML4CPPProfile::Singleton
-				static const long long SINGLETON_CLASS = 1215028894;
+				static const unsigned long SINGLETON_CLASS = 1215028894;
 				static const unsigned int SINGLETON_CLASS_FEATURE_COUNT = 1;
 				static const unsigned int SINGLETON_CLASS_OPERATION_COUNT = 0;
 				//Class UML4CPPProfile::UML4CPPPackage
-				static const long long UML4CPPPACKAGE_CLASS = 1707580997;
+				static const unsigned long UML4CPPPACKAGE_CLASS = 1707580997;
 				static const unsigned int UML4CPPPACKAGE_CLASS_FEATURE_COUNT = 4;
 				static const unsigned int UML4CPPPACKAGE_CLASS_OPERATION_COUNT = 0;
