+++ src/common/UML4CPPProfile/src_gen/UML4CPPProfile/impl/MainBehaviorImpl.cpp
@@ -52,17 +52,17 @@ MainBehaviorImpl::MainBehaviorImpl()
 	// init Get Set
 	//getter init
 		//Property base_Behavior
-		m_getterMap.insert(std::pair<long long,std::function<Any()>>(1956712138,[this](){ return eAny(this->getBase_Behavior(), uml::umlPackage::BEHAVIOR_CLASS, false);}));
+		m_getterMap.insert(std::pair<unsigned long,std::function<Any()>>(1956712138,[this](){ return eAny(this->getBase_Behavior(), uml::umlPackage::BEHAVIOR_CLASS, false);}));
 	
 	
 	//setter init
 	//Property base_Behavior
-		m_setterMap.insert(std::pair<long long,std::function<void(Any)>>(1956712138,[this](Any object){this->setBase_Behavior(object->get<std::shared_ptr<uml::Behavior>>());}));
+		m_setterMap.insert(std::pair<unsigned long,std::function<void(Any)>>(1956712138,[this](Any object){this->setBase_Behavior(object->get<std::shared_ptr<uml::Behavior>>());}));
 	
 	
 	//unsetter init
 		//Property base_Behavior
-		m_unsetterMap.insert(std::pair<long long,std::function<void()>>(1956712138,[this](){m_base_Behavior = std::shared_ptr<uml::Behavior>(nullptr);}));
+		m_unsetterMap.insert(std::pair<unsigned long,std::function<void()>>(1956712138,[this](){m_base_Behavior = std::shared_ptr<uml::Behavior>(nullptr);}));
 	
 	
 	
@@ -162,13 +162,13 @@ Any MainBehaviorImpl::get(std::shared_ptr<uml::Property> _property) const
 
 Any MainBehaviorImpl::get(std::string _qualifiedName) const
 {
-	long long uID = util::Util::polynomialRollingHash(_qualifiedName);
+	unsigned long uID = util::Util::polynomialRollingHash(_qualifiedName);
     return this->get(uID);
 }
 
-Any MainBehaviorImpl::get(long long _uID) const
+Any MainBehaviorImpl::get(unsigned long _uID) const
 {
-	std::map<long long, std::function<Any()>>::const_iterator iter = m_getterMap.find(_uID);
+	std::map<unsigned long, std::function<Any()>>::const_iterator iter = m_getterMap.find(_uID);
     if(iter != m_getterMap.cend())
     {
         //invoke the getter function
@@ -187,13 +187,13 @@ void MainBehaviorImpl::set(std::shared_ptr<uml::Property> _property, Any value)
 
 void MainBehaviorImpl::set(std::string _qualifiedName, Any value)
 {
-	long long uID = util::Util::polynomialRollingHash(_qualifiedName);
+	unsigned long uID = util::Util::polynomialRollingHash(_qualifiedName);
     this->set(uID, value);
 }
 
-void MainBehaviorImpl::set(long long _uID, Any value)
+void MainBehaviorImpl::set(unsigned long _uID, Any value)
 {
-	std::map<long long, std::function<void(Any)>>::const_iterator iter = m_setterMap.find(_uID);
+	std::map<unsigned long, std::function<void(Any)>>::const_iterator iter = m_setterMap.find(_uID);
     if(iter != m_setterMap.cend())
     {
         //invoke the setter function
@@ -210,13 +210,13 @@ void MainBehaviorImpl::unset(std::shared_ptr<uml::Property> _property)
 
 void MainBehaviorImpl::unset(std::string _qualifiedName)
 {
-	long long uID = util::Util::polynomialRollingHash(_qualifiedName);
+	unsigned long uID = util::Util::polynomialRollingHash(_qualifiedName);
     this->unset(uID);
 }
 
-void MainBehaviorImpl::unset(long long _uID)
+void MainBehaviorImpl::unset(unsigned long _uID)
 {
-	std::map<long long, std::function<void()>>::const_iterator iter = m_unsetterMap.find(_uID);
+	std::map<unsigned long, std::function<void()>>::const_iterator iter = m_unsetterMap.find(_uID);
     if(iter != m_unsetterMap.cend())
     {
         //invoke the unsetter function
@@ -243,13 +243,13 @@ Any MainBehaviorImpl::invoke(std::shared_ptr<uml::Operation> _operation, std::sh
 
 Any MainBehaviorImpl::invoke(std::string _qualifiedName, std::shared_ptr<Bag<Any>> _arguments)
 {
-	long long uID = util::Util::polynomialRollingHash(_qualifiedName);
+	unsigned long uID = util::Util::polynomialRollingHash(_qualifiedName);
     return this->invoke(uID, _arguments);
 }
 
-Any MainBehaviorImpl::invoke(long long _uID, std::shared_ptr<Bag<Any>> _arguments)
+Any MainBehaviorImpl::invoke(unsigned long _uID, std::shared_ptr<Bag<Any>> _arguments)
 {
-	std::map<long long, std::function<Any(std::shared_ptr<Bag<Any>>)>>::const_iterator iter = m_invocationMap.find(_uID);
+	std::map<unsigned long, std::function<Any(std::shared_ptr<Bag<Any>>)>>::const_iterator iter = m_invocationMap.find(_uID);
     if(iter != m_invocationMap.cend())
     {
         //invoke the operation
