+++ src/common/UML4CPPProfile/src_gen/UML4CPPProfile/impl/UML4CPPPackageImpl.cpp
@@ -53,29 +53,29 @@ UML4CPPPackageImpl::UML4CPPPackageImpl()
 	// init Get Set
 	//getter init
 		//Property base_Package
-		m_getterMap.insert(std::pair<long long,std::function<Any()>>(364749133,[this](){ return eAny(this->getBase_Package(), uml::umlPackage::PACKAGE_CLASS, false);}));
+		m_getterMap.insert(std::pair<unsigned long,std::function<Any()>>(364749133,[this](){ return eAny(this->getBase_Package(), uml::umlPackage::PACKAGE_CLASS, false);}));
 		//Property eclipseURI
-		m_getterMap.insert(std::pair<long long,std::function<Any()>>(1162675821,[this](){ return eAny(this->getEclipseURI(), types::typesPackage::STRING_CLASS, false);}));
+		m_getterMap.insert(std::pair<unsigned long,std::function<Any()>>(1162675821,[this](){ return eAny(this->getEclipseURI(), types::typesPackage::STRING_CLASS, false);}));
 		//Property ignoreNamespace
-		m_getterMap.insert(std::pair<long long,std::function<Any()>>(1716152545,[this](){ return eAny(this->isIgnoreNamespace(), types::typesPackage::BOOLEAN_CLASS, false);}));
+		m_getterMap.insert(std::pair<unsigned long,std::function<Any()>>(1716152545,[this](){ return eAny(this->isIgnoreNamespace(), types::typesPackage::BOOLEAN_CLASS, false);}));
 		//Property packageOnly
-		m_getterMap.insert(std::pair<long long,std::function<Any()>>(262131715,[this](){ return eAny(this->isPackageOnly(), types::typesPackage::BOOLEAN_CLASS, false);}));
+		m_getterMap.insert(std::pair<unsigned long,std::function<Any()>>(262131715,[this](){ return eAny(this->isPackageOnly(), types::typesPackage::BOOLEAN_CLASS, false);}));
 	
 	
 	//setter init
 	//Property base_Package
-		m_setterMap.insert(std::pair<long long,std::function<void(Any)>>(364749133,[this](Any object){this->setBase_Package(object->get<std::shared_ptr<uml::Package>>());}));
+		m_setterMap.insert(std::pair<unsigned long,std::function<void(Any)>>(364749133,[this](Any object){this->setBase_Package(object->get<std::shared_ptr<uml::Package>>());}));
 	//Property eclipseURI
-		m_setterMap.insert(std::pair<long long,std::function<void(Any)>>(1162675821,[this](Any object){this->setEclipseURI(object->get<std::string>());}));
+		m_setterMap.insert(std::pair<unsigned long,std::function<void(Any)>>(1162675821,[this](Any object){this->setEclipseURI(object->get<std::string>());}));
 	//Property ignoreNamespace
-		m_setterMap.insert(std::pair<long long,std::function<void(Any)>>(1716152545,[this](Any object){this->setIgnoreNamespace(object->get<bool>());}));
+		m_setterMap.insert(std::pair<unsigned long,std::function<void(Any)>>(1716152545,[this](Any object){this->setIgnoreNamespace(object->get<bool>());}));
 	//Property packageOnly
-		m_setterMap.insert(std::pair<long long,std::function<void(Any)>>(262131715,[this](Any object){this->setPackageOnly(object->get<bool>());}));
+		m_setterMap.insert(std::pair<unsigned long,std::function<void(Any)>>(262131715,[this](Any object){this->setPackageOnly(object->get<bool>());}));
 	
 	
 	//unsetter init
 		//Property base_Package
-		m_unsetterMap.insert(std::pair<long long,std::function<void()>>(364749133,[this](){m_base_Package = std::shared_ptr<uml::Package>(nullptr);}));
+		m_unsetterMap.insert(std::pair<unsigned long,std::function<void()>>(364749133,[this](){m_base_Package = std::shared_ptr<uml::Package>(nullptr);}));
 	
 	
 	
@@ -215,13 +215,13 @@ Any UML4CPPPackageImpl::get(std::shared_ptr<uml::Property> _property) const
 
 Any UML4CPPPackageImpl::get(std::string _qualifiedName) const
 {
-	long long uID = util::Util::polynomialRollingHash(_qualifiedName);
+	unsigned long uID = util::Util::polynomialRollingHash(_qualifiedName);
     return this->get(uID);
 }
 
-Any UML4CPPPackageImpl::get(long long _uID) const
+Any UML4CPPPackageImpl::get(unsigned long _uID) const
 {
-	std::map<long long, std::function<Any()>>::const_iterator iter = m_getterMap.find(_uID);
+	std::map<unsigned long, std::function<Any()>>::const_iterator iter = m_getterMap.find(_uID);
     if(iter != m_getterMap.cend())
     {
         //invoke the getter function
@@ -240,13 +240,13 @@ void UML4CPPPackageImpl::set(std::shared_ptr<uml::Property> _property, Any value
 
 void UML4CPPPackageImpl::set(std::string _qualifiedName, Any value)
 {
-	long long uID = util::Util::polynomialRollingHash(_qualifiedName);
+	unsigned long uID = util::Util::polynomialRollingHash(_qualifiedName);
     this->set(uID, value);
 }
 
-void UML4CPPPackageImpl::set(long long _uID, Any value)
+void UML4CPPPackageImpl::set(unsigned long _uID, Any value)
 {
-	std::map<long long, std::function<void(Any)>>::const_iterator iter = m_setterMap.find(_uID);
+	std::map<unsigned long, std::function<void(Any)>>::const_iterator iter = m_setterMap.find(_uID);
     if(iter != m_setterMap.cend())
     {
         //invoke the setter function
@@ -263,13 +263,13 @@ void UML4CPPPackageImpl::unset(std::shared_ptr<uml::Property> _property)
 
 void UML4CPPPackageImpl::unset(std::string _qualifiedName)
 {
-	long long uID = util::Util::polynomialRollingHash(_qualifiedName);
+	unsigned long uID = util::Util::polynomialRollingHash(_qualifiedName);
     this->unset(uID);
 }
 
-void UML4CPPPackageImpl::unset(long long _uID)
+void UML4CPPPackageImpl::unset(unsigned long _uID)
 {
-	std::map<long long, std::function<void()>>::const_iterator iter = m_unsetterMap.find(_uID);
+	std::map<unsigned long, std::function<void()>>::const_iterator iter = m_unsetterMap.find(_uID);
     if(iter != m_unsetterMap.cend())
     {
         //invoke the unsetter function
@@ -296,13 +296,13 @@ Any UML4CPPPackageImpl::invoke(std::shared_ptr<uml::Operation> _operation, std::
 
 Any UML4CPPPackageImpl::invoke(std::string _qualifiedName, std::shared_ptr<Bag<Any>> _arguments)
 {
-	long long uID = util::Util::polynomialRollingHash(_qualifiedName);
+	unsigned long uID = util::Util::polynomialRollingHash(_qualifiedName);
     return this->invoke(uID, _arguments);
 }
 
-Any UML4CPPPackageImpl::invoke(long long _uID, std::shared_ptr<Bag<Any>> _arguments)
+Any UML4CPPPackageImpl::invoke(unsigned long _uID, std::shared_ptr<Bag<Any>> _arguments)
 {
-	std::map<long long, std::function<Any(std::shared_ptr<Bag<Any>>)>>::const_iterator iter = m_invocationMap.find(_uID);
+	std::map<unsigned long, std::function<Any(std::shared_ptr<Bag<Any>>)>>::const_iterator iter = m_invocationMap.find(_uID);
     if(iter != m_invocationMap.cend())
     {
         //invoke the operation
